Profile page refactor
