+++ resources/js/components/Profile.js
@@ -16,9 +16,6 @@ const GRID_MODE = true;
 const LIST_MODE = false;
 
 const CreateNewSet = (name) => {
-    //console.log("New set name: " + name);
-
-
     //send new set to database and get id
     //redirect("/set", [{key: "id", value: newsetid}]);
 }
@@ -59,25 +56,10 @@ const SetsDisplay = ({displayMode, setInfo}) => {
 }
 
 const Profile = ({userInfo, userSets, currentUser}) => {
-    const [editing, setEditing] = useState(false);
-    const [displayType, setDisplayType] = useState(GRID_MODE);
-
-    const [modalOpen, setModalOpen] = useState(false);
     const [newSetName, setNewSetName] = useState("");
-
-    const [image, setImage] = useState(null);
-
-    const [imageid, setImageid] = useState((userInfo && userInfo.imageid) ? userInfo.imageid : tmp_pic);
-    const [name, setName] = useState((userInfo && userInfo.username) ? userInfo.username : "");
-    const [bio, setBio] = useState((userInfo && userInfo.bio) ? userInfo.bio : "");
-
-    useEffect(() => {
-        if (!userInfo) return;
-        setName(userInfo.username);
-        setImageid(userInfo.imageid);
-        setBio(userInfo.bio);
-    }, [userInfo]);
-
+    const [displayType, setDisplayType] = useState(GRID_MODE);
+    const [modalOpen, setModalOpen] = useState(false);
+    
     const copyLinkToProfile = () => {
         let temp = document.createElement('input');
         let linkToProfile = window.location.href;
@@ -91,103 +73,21 @@ const Profile = ({userInfo, userSets, currentUser}) => {
         toast("Copied profile link to clipboard!", "success");
     }
 
-    const onImageInput = (e) => {
-        //change image locally
-        setImage(e.target.files[0]);
-        console.log("Image input event: ", e.target.files[0]);
-        let pfp_file = document.getElementById("pfp-upload").files;
-        let imgObj;
-        if (pfp_file && pfp_file[0]) imgObj = URL.createObjectURL(pfp_file[0]);
-
-        setImageid(imgObj);
-    }
-
-    const onSubmit = async () => {
-        let updatedName = document.getElementById("username-input").value;
-        updatedName = updatedName === "" ? userInfo.username : updatedName;
-        let updatedBio = document.getElementById("description-textarea").value;
-        await axios.put(`/api/user/`, {
-            userid: currentUser.userid,
-            username: updatedName, 
-            bio: updatedBio
-        });
-
-        //create form data for image change in database
-        if (image){
-            var formData = new FormData();
-            formData.append("image", image);  
-            uploadFile(`/api/user/${currentUser.userid}/image`, formData);
-        }
-
-        setEditing(false);
-
-        redirect(window.location.href);
-    }
-
-    const onCancel = () => {
-        setImageid(userInfo.imageid);
-        setEditing(false);
-    }
-
-    const onCreateNewSet = () => {
-
-    }
-
-    const ProfileDisplay = ({editingMode}) => {
-        if (!editingMode){
-            return (
-                <div>
-                    <Image size = "small" src = {imageid} circular centered/>
-                    <Header as = "h1" textAlign = "center" >{name}</Header>
-                    <Segment>
-                        <Header as = "h3" textAlign = "center" >{bio}</Header>
-                    </Segment>
-                    { currentUser && userInfo && currentUser.userid === userInfo.userid &&
-                    <div style = {{width: "100%", textAlign: "center", marginTop: "60px"}}>
-                        <Button icon onClick = {() => {
-                            setEditing(true);
-                            setName(userInfo.username);
-                            setImageid(userInfo.imageid);
-                            setBio(userInfo.bio);
-                        }}><Icon name = "pencil"/></Button>
-                        <Button icon onClick = {() => copyLinkToProfile()}><Icon name = "linkify"/></Button>
-                    </div> }
-                </div>
-            );
-        }
-        else{
-            return(
-                <div>
-                    <Image size = "small" src = {imageid} circular centered/>
-                    
-                    <Form role = "form" encType = "multipart/form-data">
-                        <div style = {{width: "100%", textAlign: "center", marginTop: "30px"}}>
-                            <Button className = "pfp-upload-button" id = "pfp-upload-button"><label htmlFor = "pfp-upload" className = "pfp-upload-label">Upload</label></Button>
-                            <input id = "pfp-upload" onChange = {onImageInput} hidden type = "file" accept = ".jpg, .jpeg, .png"/>
-                        </div>
-    
-                        <Input fluid className = "username-input" id = "username-input" size = "large" defaultValue = {name} />
-
-                        <Segment>
-                            <TextArea id = "description-textarea" style = {{resize: "vertical"}} defaultValue = {bio}></TextArea>
-                        </Segment>
-
-                        <div style = {{width: "100%", textAlign: "center", marginTop: "60px"}}>
-                            <Button className = "profile-save-button" color = "green" onClick = {onSubmit}>Save</Button>
-                            <Button color = "red" onClick = {onCancel}>Cancel</Button>
-                        </div>
-                    </Form>
-                </div>
-            );
-        }
-        
-    }
-
     return (
         <Grid divided padded stackable columns = {2} >
             <Grid.Column width = {5}>
                 <Segment padded = "very">
-                    <ProfileDisplay editingMode = {editing} />
+                    { userInfo && <div>
+                        <Image size = "small" src = {userInfo.imageid} circular centered/>
+                        <Header as = "h1" textAlign = "center" >{userInfo.username}</Header>
+                        <Segment>
+                            <Header as = "h3" textAlign = "center" >{userInfo.bio}</Header>
+                        </Segment>
+                        { currentUser && userInfo && currentUser.userid === userInfo.userid &&
+                        <div style = {{width: "100%", textAlign: "center", marginTop: "60px"}}>
+                            <Button icon onClick = {() => copyLinkToProfile()}><Icon name = "linkify"/></Button>
+                        </div> }
+                    </div>}
                 </Segment>
             </Grid.Column>
 
@@ -202,7 +102,7 @@ const Profile = ({userInfo, userSets, currentUser}) => {
                             onClose = {() => setModalOpen(false)}
                             onOpen = {() => setModalOpen(true)}
                             open = {modalOpen}
-                            trigger = {<Button floated = "left" icon onClick = {() => onCreateNewSet()} ><Icon name = "plus"/></Button>}>
+                            trigger = {<Button floated = "left" icon ><Icon name = "plus"/></Button>}>
 
                                 <Modal.Header>Enter your set's name</Modal.Header>
                                 <Modal.Content>
